+++ Tkl.Jumbo.Jet/TaskExecutionUtility.cs
@@ -536,7 +536,7 @@ namespace Tkl.Jumbo.Jet
                 else
                 {
                     Type multiInputRecordReaderType = Configuration.StageConfiguration.MultiInputRecordReaderType.ReferencedType;
-                    int bufferSize = multiInputRecordReaderType == typeof(MergeRecordReader<T>) ? JetClient.Configuration.FileChannel.MergeTaskReadBufferSize : JetClient.Configuration.FileChannel.ReadBufferSize;
+                    int bufferSize = multiInputRecordReaderType == typeof(MergeRecordReader<T>) ? (int)JetClient.Configuration.FileChannel.MergeTaskReadBufferSize : (int)JetClient.Configuration.FileChannel.ReadBufferSize;
                     CompressionType compressionType = Configuration.JobConfiguration.GetTypedSetting(FileOutputChannel.CompressionTypeSetting, JetClient.Configuration.FileChannel.CompressionType);
                     MultiInputRecordReader<T> reader = (MultiInputRecordReader<T>)JetActivator.CreateInstance(multiInputRecordReaderType, this, InputChannels.Count, AllowRecordReuse, bufferSize, compressionType);
                     foreach( IInputChannel inputChannel in InputChannels )
